plugins/behaviors: charset exception during html processing
It seems to be possible that some html+js scripts change/send chars not
specified on the html original encoding, breaking our current text handler that
tries to find website's title.

Because of that we need to catch the exception and try to find the title of the
website only within the "correct charset" range.