hook: pre-commit hook for flake8 compliance
A pre-commit hook to check flake8 standards in the changed code. It helps
preventing too many cycles between CI and developer to check && correct coding
style issues something. In this way we guarantee the code is following the
coding style rules before it even reach the remote source tree.

But it's important to remember to run 'pre-commit install' before you continue
your work after this commit or whenever you clone the project git repo. The
first time the hook runs it might take some time to finish (few seconds), but
it'll be true only for the first time.